EES-4781 remove data catalogue feature flag
