+++ src/explore-education-statistics-admin/src/pages/release/pre-release/__tests__/PreReleaseContentPage.test.tsx
@@ -93,7 +93,7 @@ describe('PreReleaseContentPage', () => {
       slug: 'publication-1',
       releases: [],
       releaseSeries: [],
-      topic: { theme: { title: 'Theme 1' } },
+      topic: { theme: { id: 'theme-1', title: 'Theme 1' } },
       contact: {
         contactName: 'John Smith',
         contactTelNo: '0777777777',
