+++ src/explore-education-statistics-frontend/src/components/SortControls.tsx
@@ -1,9 +1,9 @@
 import Button from '@common/components/Button';
 import { FormRadioGroup, FormSelect } from '@common/components/form';
 import { useMobileMedia } from '@common/hooks/useMedia';
-import { PublicationSortOption } from '@common/services/publicationService';
 import styles from '@frontend/components/SortControls.module.scss';
-import { DataSetFileSortOption } from '@frontend/services/dataSetFileService';
+import { DataSetFileSortOption } from '@frontend/modules/data-catalogue/utils/dataSetFileSortOptions';
+import { PublicationSortOption } from '@frontend/modules/find-statistics/utils/publicationSortOptions';
 import classNames from 'classnames';
 import React from 'react';
