+++ src/explore-education-statistics-frontend/src/modules/data-catalogue/DataSetFilePage.tsx
@@ -129,7 +129,7 @@ export default function DataSetFilePage({
   return (
     <Page
       title={title}
-      caption={`Data set from ${release.publication.themeTitle}`}
+      caption={`Data set from ${release.publication.title}`}
       breadcrumbs={[{ name: 'Data catalogue', link: '/data-catalogue' }]}
       wide={fullScreenPreview}
     >
