+++ src/explore-education-statistics-frontend/src/modules/data-catalogue/utils/dataSetFileFilters.ts
@@ -1,0 +1,10 @@
+export const dataSetFileFilters = [
+  'dataSetType',
+  'latest',
+  'publicationId',
+  'releaseId',
+  'searchTerm',
+  'themeId',
+] as const;
+
+export type DataSetFileFilter = (typeof dataSetFileFilters)[number];
