+++ src/explore-education-statistics-frontend/src/modules/data-catalogue/utils/dataSetFileSortOptions.ts
@@ -1,0 +1,8 @@
+export const dataSetFileSortOptions = [
+  'newest',
+  'oldest',
+  'relevance',
+  'title',
+] as const;
+
+export type DataSetFileSortOption = (typeof dataSetFileSortOptions)[number];
