+++ src/explore-education-statistics-frontend/src/modules/find-statistics/PublicationReleasePage.tsx
@@ -450,7 +450,7 @@ const PublicationReleasePage: NextPage<Props> = ({ release }) => {
           }
           renderDataCatalogueLink={
             <Link
-              to={`/data-catalogue/${release.publication.slug}/${release.slug}`}
+              to={`/data-catalogue?themeId=${release.publication.topic.theme.id}&publicationId=${release.publication.id}&releaseId=${release.id}`}
             >
               Data catalogue
             </Link>
