+++ src/explore-education-statistics-frontend/src/modules/find-statistics/__tests__/__data__/testReleaseData.ts
@@ -53,6 +53,7 @@ export const testPublication: Publication = {
   ],
   topic: {
     theme: {
+      id: 'test-theme',
       title: 'Pupils and schools',
     },
   },
