+++ src/explore-education-statistics-frontend/src/modules/find-statistics/utils/publicationFilters.ts
@@ -1,0 +1,3 @@
+export const publicationFilters = ['releaseType', 'search', 'themeId'] as const;
+
+export type PublicationFilter = (typeof publicationFilters)[number];
