+++ src/explore-education-statistics-frontend/src/modules/find-statistics/utils/publicationSortOptions.ts
@@ -1,0 +1,8 @@
+export const publicationSortOptions = [
+  'newest',
+  'oldest',
+  'relevance',
+  'title',
+] as const;
+
+export type PublicationSortOption = (typeof publicationSortOptions)[number];
