+++ src/explore-education-statistics-frontend/src/modules/table-tool/components/__tests__/__data__/tableData.ts
@@ -278,6 +278,7 @@ export const testPublicationRelease: Release = {
     topic: {
       theme: {
         title: '',
+        id: '',
       },
     },
     contact: {
